# Swagger_3.8.8.yml
openapi: 3.0.1
info:
  title: openapiproblems
  version: 0.0.1
paths:
  /api/email/output:
    get:
      tags:
      - Email
      summary: Get the email output location.
      description: Get the email output location.
      operationId: getEmailOutputLocation
      responses:
        "200":
          description: Email output location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadEmailOutputLocationDto10'
        "401":
          description: Not authenticated
        "500":
          description: Internal server error
  /api/email/settings:
    get:
      tags:
      - Email
      summary: Get the email settings.
      description: Get the email settings.
      operationId: getEmailSettings
      responses:
        "200":
          description: Email settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadEmailSettingsDto10'
        "401":
          description: Not authenticated
        "500":
          description: Internal server error
...
components:
  schemas:
...
    EmailSendProtocolDto10:
      type: string
      description: Schema that represents the possible email protocols for sending
        emails
      enum:
      - SMTP
      - SMTP_SSL
      - SMTP_STARTTLS
...
    ReadEmailOutputLocationDto10:
      required:
      - protocol
      type: object
      properties:
        protocol:
          allOf:
          - $ref: '#/components/schemas/EmailSendProtocolDto10'
          - type: object
            description: Protocol used for the connection
      description: Schema that represents an existing email output location
    ReadEmailSettingsDto10:
      required:
      - active
      - hostname
      - plaintextPassword
      - port
      - protocol
      - senderEmail
      - username
      type: object
      properties:
        active:
          type: boolean
          description: "Flag that indicates whether the email sending is active or\
            \ not. If set to false, all other values are null"
        hostname:
          type: string
          description: Hostname or IP of the email server or null if email sending
            is disabled
          nullable: true
        port:
          type: integer
          description: Port of the email server or null if email sending is disabled
          format: int32
          nullable: true
        protocol:
          nullable: true
          allOf:
          - $ref: '#/components/schemas/EmailSendProtocolDto10'
          - type: object
            description: Protocol used for the connection or null if email sending
              is disabled
        username:
          type: string
          description: Email username to login or null if email sending is disabled
          nullable: true
        plaintextPassword:
          type: string
          description: Plaintext password for the email user to login in or null if
            email sending is disabled
          nullable: true
        senderEmail:
          type: string
          description: Sender email address that is used to send emails or null if
            email sending is disabled
          nullable: true
      description: Schema that represents the current email settings
...